split graphQL generation code, allow schema reset and fix tests
