+++ packages/vulcan-lib/lib/modules/graphql/index.js
@@ -1,0 +1,2 @@
+export * from './graphql';
+export * from './typedefs';
