+++ packages/vulcan-lib/lib/modules/index.js
@@ -11,7 +11,7 @@ import './icons.js';
 export * from './components.js';
 export * from './collections.js';
 export * from './callbacks.js';
-export * from './graphql.js';
+export * from './graphql';
 export * from './routes.js';
 export * from './utils.js';
 export * from './settings.js';
