+++ packages/vulcan-lib/lib/server/apollo-server/initGraphQL.js
@@ -4,72 +4,13 @@
 
 import { makeExecutableSchema } from 'apollo-server';
 import { mergeSchemas } from 'graphql-tools';
-import { GraphQLSchema } from '../../modules/graphql.js';
+import { GraphQLSchema, generateTypeDefs } from '../../modules/graphql';
 import { runCallbacks } from '../../modules/callbacks.js';
 
-const getQueries = () =>
-  `type Query {
-${GraphQLSchema.queries
-    .map(
-      q =>
-        `${
-          q.description
-            ? `  # ${q.description}
-`
-            : ''
-        }  ${q.query}
-  `
-    )
-    .join('\n')}
-}
-
-`;
-const getMutations = () =>
-  GraphQLSchema.mutations.length > 0
-    ? `
-${
-        GraphQLSchema.mutations.length > 0
-          ? `type Mutation {
-
-${GraphQLSchema.mutations
-              .map(
-                m =>
-                  `${
-                    m.description
-                      ? `  # ${m.description}
-`
-                      : ''
-                  }  ${m.mutation}
-`
-              )
-              .join('\n')}
-}
-`
-          : ''
-      }
-
-`
-    : '';
-// typeDefs
-const generateTypeDefs = () => [
-  `
-scalar JSON
-scalar Date
-
-${GraphQLSchema.getAdditionalSchemas()}
-
-${GraphQLSchema.getCollectionsSchemas()}
-
-${getQueries()}
-
-${getMutations()}
-
-`,
-];
 
 const initGraphQL = () => {
   runCallbacks('graphql.init.before');
-  const typeDefs = generateTypeDefs();
+  const typeDefs = generateTypeDefs(GraphQLSchema);
   const executableSchema = makeExecutableSchema({
     typeDefs,
     resolvers: GraphQLSchema.resolvers,
